Don't make backup if library doesn't already exist

+++ config/services/kiwix/generate-library.sh
@@ -6,5 +6,8 @@ library_file_tmp=$(mktemp)
 zims_dir="${KIWIX_DATA_DIR}/zims"
 
 docker run --user "$(id -u):$(id -g)" --rm -i -v "${KIWIX_DATA_DIR}:${KIWIX_DATA_DIR}" --name kiwix-quickstart --entrypoint bash kiwix/library -c "kiwix-manage ${library_file_tmp} add $(ls ${zims_dir} | xargs)"
-mv ${library_file} ${library_file}.$(date '+%Y-%m-%d_%H-%M-%S').bak
+if [ -e ${library_file} ]
+  then
+  mv ${library_file} ${library_file}.$(date '+%Y-%m-%d_%H-%M-%S').bak
+fi
 mv ${library_file_tmp} ${library_file}
